refactor(GameSearch): extract BGG API logic into useBoardGameSearch composable
Moves all BoardGameGeek fetch, XML parse, sort/filter, and debounce
logic out of GameSearch.vue into composables/useBoardGameSearch.ts.
The component is now a thin UI wrapper (~30 lines of script).
